exercises/md/ex3/requirements.txt
matplotlib==3.7.1
numpy==1.25.0
Pillow==11.0.0
tqdm==4.66.1
